+++ extensions/ql-vscode/src/view/results/RawTable.tsx
@@ -11,7 +11,7 @@ import {
 import RawTableHeader from "./RawTableHeader";
 import RawTableRow from "./RawTableRow";
 import { ResultRow } from "../../common/bqrs-cli-types";
-import { onNavigation } from "./results";
+import { onNavigation } from "./ResultsApp";
 import { tryGetResolvableLocation } from "../../common/bqrs-utils";
 import { ScrollIntoViewHelper } from "./scroll-into-view-helper";
 import { sendTelemetry } from "../common/telemetry";
