+++ extensions/ql-vscode/src/view/results/__tests__/results.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { render as reactRender, screen } from "@testing-library/react";
-import { ResultsApp } from "../results";
+import { ResultsApp } from "../ResultsApp";
 import {
   Interpretation,
   IntoResultsViewMsg,
