+++ extensions/ql-vscode/src/view/results/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { WebviewDefinition } from "../webview-definition";
-import { ResultsApp } from "./results";
+import { ResultsApp } from "./ResultsApp";
 
 const definition: WebviewDefinition = {
   component: <ResultsApp />,
